impl(storage): add a function for building XML for MPU
