+++ google/cloud/storage/google_cloud_cpp_storage.bzl
@@ -110,6 +110,7 @@ google_cloud_cpp_storage_hdrs = [
     "internal/tuple_filter.h",
     "internal/unified_rest_credentials.h",
     "internal/xml_builders.h",
+    "internal/xml_escape.h",
     "internal/xml_node.h",
     "lifecycle_rule.h",
     "list_buckets_reader.h",
@@ -226,6 +227,7 @@ google_cloud_cpp_storage_srcs = [
     "internal/signed_url_requests.cc",
     "internal/unified_rest_credentials.cc",
     "internal/xml_builders.cc",
+    "internal/xml_escape.cc",
     "internal/xml_node.cc",
     "lifecycle_rule.cc",
     "list_buckets_reader.cc",
