+++ google/cloud/storage/internal/xml_escape_test.cc
@@ -12,34 +12,44 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_STORAGE_INTERNAL_XML_BUILDERS_H
-#define GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_STORAGE_INTERNAL_XML_BUILDERS_H
-
 #include "google/cloud/storage/internal/xml_escape.h"
-#include "google/cloud/storage/internal/xml_node.h"
-#include "google/cloud/storage/version.h"
-#include "google/cloud/status_or.h"
-#include <map>
-#include <memory>
-#include <mutex>
-#include <string>
+#include <gmock/gmock.h>
 
 namespace google {
 namespace cloud {
 namespace storage_internal {
 GOOGLE_CLOUD_CPP_INLINE_NAMESPACE_BEGIN
+namespace {
 
-/**
- * A function for creating an XML request for `Completing multipart upload` API
- * described at
- * https://cloud.google.com/storage/docs/xml-api/post-object-complete.
- */
-StatusOr<std::shared_ptr<XmlNode>> BuildCompleteMultipartUploadXml(
-    std::map<unsigned int, std::string> const& parts);
 
+TEST(XmlNodeTest, EscapeXmlTag) {
+  std::vector<std::pair<std::string, std::string>> cases = {
+      {"\"Google\"", "&quot;Google&quot;"},
+      {"'Google'", "&apos;Google&apos;"},
+      {"<html>", "&lt;html&gt;"},
+      {"Bread&Butter", "Bread&amp;Butter"},
+  };
+  for (auto& p : cases) {
+    auto after = EscapeXmlTag(p.first);
+    EXPECT_EQ(after, p.second);
+  }
+}
+
+TEST(XmlNodeTest, EscapeXmlContent) {
+  std::vector<std::pair<std::string, std::string>> cases = {
+      {"\"Google\"", "\"Google\""},
+      {"'Google'", "'Google'"},
+      {"<html>", "&lt;html&gt;"},
+      {"Bread&Butter", "Bread&amp;Butter"},
+  };
+  for (auto& p : cases) {
+    auto after = EscapeXmlContent(p.first);
+    EXPECT_EQ(after, p.second);
+  }
+}
+
+}  // namespace
 GOOGLE_CLOUD_CPP_INLINE_NAMESPACE_END
 }  // namespace storage_internal
 }  // namespace cloud
 }  // namespace google
-
-#endif  // GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_STORAGE_INTERNAL_XML_BUILDERS_H
