+++ google/cloud/storage/internal/xml_node.h
@@ -25,10 +25,6 @@ namespace cloud {
 namespace storage_internal {
 GOOGLE_CLOUD_CPP_INLINE_NAMESPACE_BEGIN
 
-std::string EscapeXmlTag(std::string const& val);
-
-std::string EscapeXmlContent(std::string const& val);
-
 /**
  * Represents an XML node in an XML tree.
  *
